Make the related zendesk tickets ajax
