--- app/views/issues/_related_zendesk_tickets.html.erb
@@ -1,36 +1,0 @@
-<hr />
-<p>
-  <strong>Related Zendesk Tickets (<%= @issue.related_zendesk_tickets.count %>)</strong>
-</p>
-<% if @issue.related_zendesk_tickets.count > 0 %>
-<table class="list issues">
-  <thead>
-    <tr>
-      <th>Zendesk ID</th>
-      <th>Status</th>
-      <th>Subject</th>
-      <th>Assigned</th>
-      <th>Created</th>
-      <th>Updated</th>
-    </tr>
-  </thead>
-  <tbody>
-    <% @issue.related_zendesk_tickets.each do |ticket| %>
-    <tr>
-      <td class="id"><%= link_to ticket[:id], "https://#{@issue.zendesk_subdomain}.zendesk.com/agent/tickets/#{ticket[:id]}", target: '_blank' %>
-      <td class="status"><%= ticket[:status] %></td>
-      <td class="subject"><%= link_to ticket[:subject], "https://#{@issue.zendesk_subdomain}.zendesk.com/agent/tickets/#{ticket[:id]}", target: '_blank' %></td>
-      <% if ticket.assignee.nil? %>
-      <td class="assignee_id"><%= ticket.group.name %></td>
-      <% else %>
-      <td class="assignee_id"><%= mail_to ticket.assignee.email, ticket.assignee.name, title: ticket.group.name %></td>
-      <% end %>
-      <td class="created_at"><span title="<%= ticket[:created_at] %>"><%= ticket[:created_at].strftime('%m/%d/%Y') %></span></td>
-      <td class="updated_at"><span title="<%= ticket[:updated_at] %>"><%= ticket[:updated_at].strftime('%m/%d/%Y') %></span></td>
-    </tr>
-    <% end %>
-  </body>
-</table>
-<% else %>
-<p>No related tickets.</p>
-<% end %>
